pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "xeen"
version = "0.1.19"
description = "Screenshot capture → edit → crop → publish. One command: xeen"
readme = "README.md"
license = "Apache-2.0"
requires-python = ">=3.10"
authors = [
    { name = "Tom Sapletta", email = "tom@sapletta.com" },
]
keywords = ["screenshot", "video", "capture", "social-media", "screen-recording"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Web Environment",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Multimedia :: Video :: Capture",
]
dependencies = [
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "mss>=9.0.0",
    "Pillow>=10.0.0",
    "pynput>=1.7.6",
    "jinja2>=3.1.0",
    "python-multipart>=0.0.6",
    "aiofiles>=23.0.0",
    "numpy>=1.24.0",
    "imagehash>=4.3.0",
    "pytesseract>=0.3.10",
    "loguru>=0.7.0",
    "litellm>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21",
    "httpx>=0.25",
    "ruff>=0.1.0",
    "twine>=4.0",
]
ocr = [
    "pytesseract>=0.3.10",
]
desktop = [
    "pywebview>=4.0",
]
all = [
    "pytesseract>=0.3.10",
    "pywebview>=4.0",
]

[project.scripts]
xeen = "xeen.cli:main"
xeen-capture = "xeen.cli:main"

[project.urls]
Homepage = "https://github.com/softreck/xeen-capture"
Repository = "https://github.com/softreck/xeen-capture"

[tool.setuptools.packages.find]
include = ["xeen*"]

[tool.setuptools.package-data]
xeen = ["static/**/*", "templates/**/*"]

[tool.ruff]
line-length = 100
target-version = "py310"
